Basic editing didn't work so hot
Slightly hotter now but still buggy.

+++ App/Event.h
@@ -24,8 +24,8 @@ enum class BufferEventType {
 };
 
 struct EventRange {
-    Vec<int> start {};
-    Vec<int> end {};
+    Vec<size_t> start {};
+    Vec<size_t> end {};
 };
 
 struct Buffer;
